* moved autosave session logic into createCommon * improved enumeration of FieldTypes
