+++ index.ts
@@ -6,9 +6,8 @@ import { prompt, Separator, registerPrompt } from 'inquirer'
 import autocomplete from 'inquirer-autocomplete-prompt'
 import { readInCurrentConfig } from 'app/et/configs'
 import { ensurePathExists, format, getFiles, getIdealSizeForInquirer } from 'app/helpers'
-import { cleanupEmptySessions, isCurrentSessionEmpty, saveSession, session, SESSION_DIR } from 'app/session'
-import { DIST_JOURNEY_DIR, YES, YES_OR_NO } from 'app/constants'
-import { setSessionName } from 'app/journeys/et/sessionSetName'
+import { cleanupEmptySessions, SESSION_DIR } from 'app/session'
+import { DIST_JOURNEY_DIR } from 'app/constants'
 import { Journey } from 'types/journey'
 import { resolve } from 'path'
 
@@ -101,26 +100,6 @@ function findSelectedJourney(choices: Journey[], selected: string) {
 }
 
 /**
- * Ask the user for a session name if the current session has a default name (session_TIME) and has data in it
- */
-async function conditionalAskForSessionName() {
-  const isDefaultName = session.name.match(/^session_\d+$/)
-  if (isCurrentSessionEmpty() || !isDefaultName) {
-    return
-  }
-
-  const notEmptyQuestion = `Current session (${session.name}) is not empty but has not had a name set. Would you like to do that now?`
-  const answers = await prompt([
-    { name: 'name', message: notEmptyQuestion, type: 'list', choices: YES_OR_NO }
-  ])
-
-  if (answers.name === YES) {
-    await setSessionName()
-    saveSession(session)
-  }
-}
-
-/**
  * The main program loop. Initializes program and asks questions until "Exit" is selected
  */
 async function start() {
@@ -168,9 +147,6 @@ async function start() {
       console.error(e)
       break
     }
-
-    saveSession(session)
-    await conditionalAskForSessionName()
   }
   console.log('Bye!')
 }
